requirements.txt
Pillow==9.2.0  # For handling images
Flask==2.3.3  # Web framework
Werkzeug==2.3.3  # For secure password handling
email-validator==1.3.1  # For email validation
requests==2.31.0  # For HTTP requests
pyngrok==5.2.2  # For setting up secure tunnels
python-dotenv==1.0.0  # For loading environment variables from .env files
